+++ index.js
@@ -1,0 +1,9 @@
+import { GroundControl } from "./ground-control.js";
+import { InputControl } from "./input-control.js";
+import { ToggleControl } from "./toggle-control.js";
+
+export default {
+  GroundControl,
+  InputControl,
+  ToggleControl,
+};